requirements.txt
langchain==0.1.0
groq==0.4.1
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0